Set part of main structure in HTML and style it
All of que required files were built, that is:
- The HTML file.
- The CSS file.
- The javascript file.

In the HTML the header was set and the container for the main body.
On the other hand, in the CSS file the sections were placed in the
desired place.

Likewise, within the main container 9 div elements were set, which
will be the squares for the gameboard in which the player can click.