Create responsive gameboard
In the HTML file 9 squares were created, each one wrapped in a DIV
element.

In the CSS file the container of the divs containing the squares, which
is the gameboard, was set with a width of 35% and with a display of
grid, allowing to specify a 3x3 grid. Then, the DIVS in the next level
were given a relative position and added a pseudo-element, whose display
is table and padding-bottom of 100%.

Finally, each square was given an absolute position and placed in the
upper left corner, with a height and widht of 100%.