Change layout and add sidebar, counter sections
The layout was change to improve the appearance of the webpage. The body
was set to be grid, in which a grid of 2x2 was specified. In the upper
left part is the header, below it the section for the buttons was set,
and in the upper right part the counter. The gameboard occupies the
remaining part.

The gameboard has a black background color, whereas the squares white.
Thanks to this, by giving a gap in the gameboard, the divisions for the
tic tac toe game are already being displayed.