Create factory function for players and game object
Once the object for the board game was created, it was given the method
of changing any of its squares by means of an event listener 'click';
however, it would invoke the method 'playerToPlay' defined in the game
object.

The game object contains, so far, the method previously mentioned. This
method invokes the method 'putXorO' from the players object. Morever,
the playerToPlay method leverages a variable, named currentTurn, to
decide the current turn. If its value is 1, then player 1 applies
putXorO and currentTurn increases its value. When the value is different
from 1, then player 2 makes its play and the currentTurn is set to 1
again.

The game object is created when the START button is pressed.

The players object receives as arguments the name of the player and the
symbol they are going to set on the squares of the board.

The players objects were defined manully at the moment, later a form in
html will be used to allow the players to decide the corresponding name.