Create perfect circle to be displayed on the board
In order to create the circle a DIV is appended to the corresponding
DIV that triggered the event, only when the player with the circle
symbol has played. The new DIV is appended with the respective class
that adds a white border and 50% radius to give the appearance of a
circle.