Fix bugs when the square is already occupied
The first bug came about when the div for the circle was introduced, as
the event listener took those DIVS to insert another circle. It was
fixed by adding three 'empty' classes to it as soon as it is created,
that way it not longer met the condition that the number of classes must
be less than 3, so the instruction was ignored.

On the other hand, when a square was already occupied but the user kept
on clicking on it, it was possible to make 2 consecutive x. This was
fixed by removing the variable currentTurn from the game object and
placing it withing the player object. As soon as the putXorO is called,
this variable is increased by 1, but if none of the conditions were met,
then the variable decreased by 1 to allow the current player to click on
other DIV. Likewise, when the second player made her or his move, the
variable was set to 1 again to continue the cycle.