Add victory method to game object
This method oversees whether a victory has happened or not. This was
achieved by implementing a brute approach, which is setting the whole
repertoire of DIV's combinations and wrapping its playerToPlay method
when is being called in the event listener. If there is a victory,
then the playerToPlay is skipped, simulating that the game has ended.