Modify victory method of game object
In order to improve this method, the program no longer adds a new div
to the board when the player with the circle has to play, it only adds
the corresponding class to modify the current DIV. Thus, the method of
adding several classes when the circle player played were removed, as
now the browser no longer adds another symbol to an occupied square.

If the player with an X wins, then the corresponding squares are given
an extra class to color the text with gold. Conversely, when the player
with a O wins, then the respective class is added to color the borders
of the div with gold.