Create pop up to announce the winner
A new DIV element was created in the HTML file, which will contain
the winner and the symbol. In the CSS file it was added a class to
style it and scale it, as it will have a transition comming from the
center to fill the whole screen when the victory is triggered. At the
moment this popup is called by the START button to see its behavior.
Likewise,in javascript the element is added another class, which
contains the required declarations to animate the transition.