Add functionality to popup and stop game when there's victory
There is a close button on the popup, so if the players want to see
how the board is at the end they have the option. On the other hand,
when a player wins, the respective counter increases; thus, the player
object has a new method and is called in the event listener of the
board.

Finally, in order to stop players from increasing the count when clicking
the board when there is already a winner, the variable "numberOfWins"
was declared and set to 0. When there is a victory this variables's
values increases 1, stopping the event listener of the board from
calling the game method to allow players to move.