Fix message in popup and declare draw
The popup displayed the name of the corresponding player, but the symbol
was always circle. So, the method of game 'assignWinner' now selects the
div in which the symbol is displayed, and depending on the player who
wins, that DIV receives the respective class and text content.

On the other hand, the game object has the method DRAW, which follows
the same logic as assignWinner, but the text content is different.

Likewise, the variable 'numberMoves' was defined within the game board
object, which helps to define when a draw takes place. When it reaches
9 and there is no victory, then it triggers an IF conditional in the
event listener to popup that it is a draw.