Add Restart button and functionality
In order to make this button work all of the objects were given the
method 'reset()', which takes all the variables in their scope to clean
them, both the text content and the number of classes.

When the button is pressed an event listener is triggered and calls all
of the objects and their method reset. This button also sets to 0 the
counters.