Add color to current player's counter
In order to know who is the player to make a play, the respective color
was added a green box-shadow. The class for this was declared and is
added by means of javascript. When the start button is pressed it adds
the class to player one. As the game evolves, the method 'playerToPlay'
of the game object removes and adds that class to the current player.