Link popup form to code
The input elements were linked to javascript by selecting those elements
and using their values to establish the name of each player.

The data is collected once the submit form button is pressed, which
calls the new function 'createPlayers', which prevents the default
behavior of the submit event and assigns the corresponding elements with
the respective values. Likewise it creates the player objects.