Allow machine to make a valid move
An AI will be implemented by mean of the minimax algorithm. However,
before undertaking that approach the computer must be able to play a
valid move. For instance, it can't place its move in an already
occupied square, furthermore, if there is already a winner then no
player can't continue playing until refreshing the board.

The game and player objects oversee all of this. The game object was
given an extra method, which is AI and calls a new method in the player
object, which is playWithAi. As soon as the player 1 makes a move the AI
will make its move only if there is no a draw or a winner.