Update draw method
The draw method used to check whether there was a draw or not employed
a variable that increased everytime a player made a move. However,
this could trigger the draw by clicking on a square already filled.
Thus,a list was created containing the available squares, but if
that list happens to be empty and there is no winner yet, then the
draw is triggered and the game ends.